Add layer of indirection for ExtendableIterativeTask (renamed) and add functionality to set the totalIterations (story #199)
