+++ common-lib/src/main/java/org/drugis/common/threading/AbstractIterativeComputation.java
@@ -1,8 +1,8 @@
 package org.drugis.common.threading;
 
 public abstract class AbstractIterativeComputation implements IterativeComputation {
-	private final int d_totalIterations;
-	private int d_iteration;
+	protected int d_totalIterations;
+	protected int d_iteration;
 
 	public AbstractIterativeComputation(int totalIterations) {
 		d_totalIterations = totalIterations;
