+++ common-lib/src/main/java/org/drugis/common/threading/IterativeSuspendable.java
@@ -38,8 +38,4 @@ class IterativeSuspendable extends AbstractSuspendable {
 	public void setReportingInterval(final int interval) {
 		d_interval  = interval;
 	}
-	
-	public void reset() { 
-		d_computation.reset();
-	}
 }
